refactor: colonizethis_setup wave 8 250 ratchet and logic-barrel SoT (#4624)
Lower lib/test/support file ceilings to 250, split locked-assigner search and capital mutators, and keep setup tests off the logic barrel so CI can enforce the wave-8 SoT.
