+++ SPEC/program/locked-province-assigner.md
@@ -2,7 +2,7 @@
 
 ## Scope
 
-Defines the **single** Old/New World province-painting path for current product setup: **phased** sequential growth with ranked legal neighbors, **phased feasibility pruning** (not the legacy parallel greedy island/residual pack on every future faction while only one faction grows), **depth-scoped tabu**, per-faction backtracking with **cross-faction unwind**, and **capital-generation restart** when search is stuck. **Authority:** GitHub **#1830** / **#1822** / **#1834**; this file is the in-repo contract for `packages/colonizethis_logic/lib/src/setup/locked_province_assigner.dart` and **colonizethis_data** topology gate helpers used before choosing locked vs BFS OW painting.
+Defines the **single** Old/New World province-painting path for current product setup: **phased** sequential growth with ranked legal neighbors, **phased feasibility pruning** (not the legacy parallel greedy island/residual pack on every future faction while only one faction grows), **depth-scoped tabu**, per-faction backtracking with **cross-faction unwind**, and **capital-generation restart** when search is stuck. **Authority:** GitHub **#1830** / **#1822** / **#1834**; this file is the in-repo contract for `packages/colonizethis_setup/lib/src/setup/locked_province_assigner.dart` and **colonizethis_data** topology gate helpers used before choosing locked vs BFS OW painting.
 
 There is **no** post-pass GP Old World connectivity repair. For configs **other than** the locked full-init profile, each Old World landmass still paints GPs + any minors rostered on that landmass using **`assignTerritoriesByBfsGrowth`** (multi-source growth with the same seeds and fair targets) because strict sequential locked growth is only required for the **six minors on four continents / 60 OW** delivery (#1830). **New World** tribes use **`assignTerritoriesLockedOnLandmass`** on remainder components via `_assignFactionsOnRemainderAuto`; multi-component cases use **bounded DFS** component packing (`_tryPackFactionsOntoPpComponentsDfs`). **`faction_component_bin_pack_failed`** / **`assignment_remainder_not_connected`** propagate to callers; **`runInitGame`** may **regenerate tile maps** and retry the pipeline per [game-setup-pipeline.md](game-setup-pipeline.md)—there is **no** whole-map NW **`assignTerritoriesByBfsGrowth`** fallback inside `createGameFromGeneratedMaps` for locked failures. **NW Tribe sea-bound ownership gate (locked full-init profile):** after NW painting, `createGameFromGeneratedMaps` requires **every Tribe to own ≥1 sea-bound (P–S) province** (parallel to the GP sea-bound seed rule). A violation throws **`SetupTopologyDataException`** with retriable code **`tribe_missing_sea_bound_province`** so the bounded regen loop retries with a bumped map seed (see [game-setup.md](../game/game-setup.md) § Tribe Assignment). The Tribe capital remains "sea-bound not required"; only ownership of ≥1 sea-bound province is required.
 
@@ -35,7 +35,7 @@ If search still fails: **`StateError`** from assigner; orchestration maps assign
 
 ## Tests
 
-- **AC-14 / AC-15:** `packages/colonizethis_logic/test/locked_assigner_mechanics_test.dart` — five-province hand-crafted topology (`a–m–{d,g}`, `d–w`) with mandatory seed `a`, targets `A:3` / `B:2`, ranked neighbor choice at depth **2** (stack length after `a` and `m`) so the first tried expansion is infeasible for `B` and the assigner performs **exactly one** undo (`LockedAssignerObservation.backtracks == 1`). **AC-14** asserts that map plus `backtracks >= 1` (golden `1`). **AC-15** runs the same call twice and asserts identical owner maps, identical `backtracks`, identical `capitalRestarts` (fixture expects `0`), and `backtracks < 50` (non-thrash guard), so duplicate runs exercise the same DFS / tabu bookkeeping surface as the single-run AC-14 test.
+- **AC-14 / AC-15:** `packages/colonizethis_setup/test/locked_assigner_mechanics_test.dart` — five-province hand-crafted topology (`a–m–{d,g}`, `d–w`) with mandatory seed `a`, targets `A:3` / `B:2`, ranked neighbor choice at depth **2** (stack length after `a` and `m`) so the first tried expansion is infeasible for `B` and the assigner performs **exactly one** undo (`LockedAssignerObservation.backtracks == 1`). **AC-14** asserts that map plus `backtracks >= 1` (golden `1`). **AC-15** runs the same call twice and asserts identical owner maps, identical `backtracks`, identical `capitalRestarts` (fixture expects `0`), and `backtracks < 50` (non-thrash guard), so duplicate runs exercise the same DFS / tabu bookkeeping surface as the single-run AC-14 test.
 - **AC-11 / AC-12:** `init_game_orchestrator_test.dart` — twenty fixed seeds for locked profile; **AC-12** repeats full setup twice using **one of those same seeds** (currently `17011`) and asserts identical Old World `province id → owner id` strings. When a run’s OW **and** NW topologies both match the locked multisets, tests assert **AC-1–AC-9** predicates (province quotas, P–P connectivity, sea-bound seeds, continent role counts, NW tribe layout); when multisets do not match procedural output, those ACs are out of scope for that run until generator gates land.
 
 ## Shared helpers
