+++ packages/colonizethis_setup/lib/src/setup/locked_province_assigner_types.dart
@@ -14,3 +14,22 @@ final class LockedAssignerObservation {
   int backtracks = 0;
   int capitalRestarts = 0;
 }
+
+/// Tabu key: capital generation, choice depth, province id.
+typedef LockedAssignerTabuKey = (
+  int capitalGeneration,
+  int choiceDepth,
+  String provinceId,
+);
+
+/// One DFS stack frame.
+typedef LockedAssignerPlacement = ({
+  String faction,
+  String province,
+  bool lockedSeed,
+});
+
+/// DFS return codes for the outer search loop.
+const int lockedAssignerDfsOk = 0;
+const int lockedAssignerDfsDeadEnd = 1;
+const int lockedAssignerDfsBudget = 2;
