+++ packages/colonizethis_setup/test/setup/game_setup_home_fleet_merge_test.dart
@@ -1,8 +1,9 @@
 import 'package:colonizethis_test/game_test_fixtures.dart';
 import 'package:colonizethis_test/test.dart';
 import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_logic/colonizethis_logic.dart';
+import 'package:colonizethis_setup/colonizethis_setup.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
+import 'package:colonizethis_world/colonizethis_world.dart';
 
 import 'init_game_orchestrator_test_support.dart';
 
@@ -41,9 +42,7 @@ void main() {
             ownerId: 'gp1',
             regionId: kRegionOldWorld,
             inPortAtProvinceId: 'oldWorld|p1',
-            ships: [
-              ShipInstance(id: 'ship_1', typeId: shipTypeId),
-            ],
+            ships: [ShipInstance(id: 'ship_1', typeId: shipTypeId)],
           ),
         ],
         players: [
