+++ packages/colonizethis_setup/test/setup/support/province_assignment_more_scenarios.dart
@@ -1,0 +1,227 @@
+// Scenario table densify (Refs #4349 Slice C).
+
+import 'dart:math';
+import 'package:colonizethis_test/test.dart';
+import 'package:colonizethis_setup/colonizethis_setup.dart';
+import 'scenario_runner.dart';
+
+List<RunnableScenario> provinceAssignmentScenariosMore() => [
+  rs('returns empty when no candidates available', () {
+    final available = <String>{};
+    final seeds = pickSimpleSeeds(
+      factionIds: ['f1'],
+      candidateIds: ['p1', 'p2'],
+      available: available,
+    );
+    expect(seeds, isEmpty);
+  }),
+  rs('stops when candidates exhausted', () {
+    final available = {'p1'};
+    final seeds = pickSimpleSeeds(
+      factionIds: ['f1', 'f2', 'f3'],
+      candidateIds: ['p1'],
+      available: available,
+    );
+    expect(seeds.length, 1);
+    expect(seeds.containsKey('p1'), isTrue);
+  }),
+  rs('assigns contiguous territories via BFS', () {
+    // Linear graph: p1 - p2 - p3 - p4 - p5
+    final neighbours = <String, Set<String>>{
+      'p1': {'p2'},
+      'p2': {'p1', 'p3'},
+      'p3': {'p2', 'p4'},
+      'p4': {'p3', 'p5'},
+      'p5': {'p4'},
+    };
+    final available = {'p1', 'p2', 'p3', 'p4', 'p5'};
+    final seeds = {'p1': 'f1', 'p5': 'f2'};
+    final targets = {'f1': 2, 'f2': 2};
+
+    final owners = assignTerritoriesByBfsGrowth(
+      neighbours: neighbours,
+      factionIds: ['f1', 'f2'],
+      seeds: seeds,
+      targetPerFaction: targets,
+      available: available,
+    );
+
+    expect(owners['p1'], 'f1');
+    expect(owners['p5'], 'f2');
+    // Each should have grown from their seed
+    expect(
+      owners.values.where((v) => v == 'f1').length,
+      greaterThanOrEqualTo(2),
+    );
+    expect(
+      owners.values.where((v) => v == 'f2').length,
+      greaterThanOrEqualTo(2),
+    );
+  }),
+  rs('respects maxTotal cap', () {
+    final neighbours = <String, Set<String>>{
+      'p1': {'p2', 'p3'},
+      'p2': {'p1', 'p4'},
+      'p3': {'p1', 'p5'},
+      'p4': {'p2'},
+      'p5': {'p3'},
+    };
+    final available = {'p1', 'p2', 'p3', 'p4', 'p5'};
+    final seeds = {'p1': 'f1'};
+    final targets = {'f1': 10};
+
+    final owners = assignTerritoriesByBfsGrowth(
+      neighbours: neighbours,
+      factionIds: ['f1'],
+      seeds: seeds,
+      targetPerFaction: targets,
+      available: available,
+      maxTotal: 3,
+    );
+
+    expect(owners.length, 3);
+    expect(owners['p1'], 'f1');
+  }),
+  rs('respects landmass constraint', () {
+    // p1(island 1) - p2(island 1), p3(island 2) - p4(island 2)
+    // p2 and p3 are neighbours across landmasses
+    final neighbours = <String, Set<String>>{
+      'p1': {'p2'},
+      'p2': {'p1', 'p3'},
+      'p3': {'p2', 'p4'},
+      'p4': {'p3'},
+    };
+    final landmass = {'p1': 1, 'p2': 1, 'p3': 2, 'p4': 2};
+    final available = {'p1', 'p2', 'p3', 'p4'};
+    final seeds = {'p1': 'f1', 'p4': 'f2'};
+    final targets = {'f1': 3, 'f2': 3};
+
+    final owners = assignTerritoriesByBfsGrowth(
+      neighbours: neighbours,
+      landmassIds: landmass,
+      factionIds: ['f1', 'f2'],
+      seeds: seeds,
+      targetPerFaction: targets,
+      available: available,
+    );
+
+    expect(owners['p1'], 'f1');
+    expect(owners['p2'], 'f1');
+    expect(owners['p4'], 'f2');
+    expect(owners['p3'], 'f2');
+  }),
+  rs('balances growth across factions', () {
+    // Star graph: center p0 connected to p1..p6
+    final neighbours = <String, Set<String>>{
+      'p0': {'p1', 'p2', 'p3', 'p4', 'p5', 'p6'},
+      'p1': {'p0'},
+      'p2': {'p0'},
+      'p3': {'p0'},
+      'p4': {'p0'},
+      'p5': {'p0'},
+      'p6': {'p0'},
+    };
+    final available = {'p0', 'p1', 'p2', 'p3', 'p4', 'p5', 'p6'};
+    final seeds = {'p1': 'f1', 'p2': 'f2'};
+    final targets = {'f1': 3, 'f2': 3};
+
+    final owners = assignTerritoriesByBfsGrowth(
+      neighbours: neighbours,
+      factionIds: ['f1', 'f2'],
+      seeds: seeds,
+      targetPerFaction: targets,
+      available: available,
+    );
+
+    final f1Count = owners.values.where((v) => v == 'f1').length;
+    final f2Count = owners.values.where((v) => v == 'f2').length;
+    expect(f1Count, greaterThanOrEqualTo(3));
+    expect(f2Count, greaterThanOrEqualTo(3));
+  }),
+  rs('greedy leftover assigns remaining provinces', () {
+    final neighbours = <String, Set<String>>{
+      'p1': {'p2'},
+      'p2': {'p1', 'p3'},
+      'p3': {'p2'},
+    };
+    final available = {'p1', 'p2', 'p3'};
+    final seeds = {'p1': 'f1'};
+    final targets = {'f1': 1};
+
+    final owners = assignTerritoriesByBfsGrowth(
+      neighbours: neighbours,
+      factionIds: ['f1'],
+      seeds: seeds,
+      targetPerFaction: targets,
+      available: available,
+    );
+
+    expect(owners.length, 3);
+  }),
+  rs(
+    'greedy leftover uses factionLandmassIds when picking claiming faction',
+    () {
+      final neighbours = <String, Set<String>>{
+        'a1': {'a2'},
+        'a2': {'a1'},
+        'b1': {'b2'},
+        'b2': {'b1'},
+      };
+      final landmass = {'a1': 1, 'a2': 1, 'b1': 2, 'b2': 2};
+      final available = {'a1', 'a2', 'b1', 'b2'};
+      final owners = assignTerritoriesByBfsGrowth(
+        neighbours: neighbours,
+        landmassIds: landmass,
+        factionLandmassIds: const {'f1': 1, 'f2': 2},
+        factionIds: const ['f1', 'f2'],
+        seeds: const {'a1': 'f1', 'b1': 'f2'},
+        targetPerFaction: const {'f1': 1, 'f2': 1},
+        available: available,
+        neighborShuffleRandom: Random(1),
+      );
+
+      expect(owners.length, 4);
+      expect(owners['a1'], 'f1');
+      expect(owners['b1'], 'f2');
+    },
+  ),
+  rs('greedy leftover falls back to lowest-count faction on landmass when '
+      'no touching owner exists', () {
+    final neighbours = <String, Set<String>>{
+      'a1': <String>{},
+      'c': <String>{},
+      'b1': {'b2'},
+      'b2': {'b1'},
+    };
+    final landmass = {'a1': 1, 'c': 1, 'b1': 2, 'b2': 2};
+    final available = {'a1', 'c', 'b1', 'b2'};
+    final owners = assignTerritoriesByBfsGrowth(
+      neighbours: neighbours,
+      landmassIds: landmass,
+      factionLandmassIds: const {'f1': 1, 'f2': 2},
+      factionIds: const ['f1', 'f2'],
+      seeds: const {'a1': 'f1', 'b1': 'f2'},
+      targetPerFaction: const {'f1': 1, 'f2': 1},
+      available: available,
+    );
+
+    expect(owners['c'], 'f1');
+  }),
+  rs('shuffles neighbor expansion order when Random is provided', () {
+    final neighbours = <String, Set<String>>{
+      'p0': {'p1', 'p2', 'p3'},
+      'p1': {'p0'},
+      'p2': {'p0'},
+      'p3': {'p0'},
+    };
+    final available = {'p0', 'p1', 'p2', 'p3'};
+    assignTerritoriesByBfsGrowth(
+      neighbours: neighbours,
+      factionIds: const ['f1'],
+      seeds: const {'p0': 'f1'},
+      targetPerFaction: const {'f1': 4},
+      available: available,
+      neighborShuffleRandom: Random(0),
+    );
+  }),
+];
